requirements.txt
asgiref==3.4.1
Django==3.2.7
django-taggit==1.5.1
django-tinymce==3.3.0
gunicorn==20.1.0
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
